+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Unreleased
 
+## Version 2.30.0 (2026-07-21)
+
+This release targets the OpenTelemetry SDK 1.64.0.
+
+Note that many artifacts have the `-alpha` suffix attached to their version
+number, reflecting that they will continue to have breaking changes. Please see
+[VERSIONING.md](https://github.com/open-telemetry/opentelemetry-java-instrumentation/blob/main/VERSIONING.md#opentelemetry-java-instrumentation-versioning)
+for more details.
+
 ### ⚠️ Breaking changes to non-stable APIs
 
 - Remove `ExperimentalInstrumentationModule#getModuleGroup()`. Built-in invokedynamic modules now